WIP: Correctly allocate buffer space new location
Previously, if a closure had nothing bound, then we would
not have enough space to store an info table and a slot
for its location after evacuation. When we evacuate a closure,
we change its infotable to point to the "already evacuated evac table",
which reads 8 bytes after base to the new location of the closure. We
need enough space for those bytes.

The solution is to reserve this space if we aren't using at least 8
bytes of stuff after the info table already.