Save temporary pointers before allocation
Previously, we might end up with stale heap addresses if a garbage
collection happens, and we've saved heap addresses into C variables.
This "fixes" that problem by always saving all temp pointers in that
function to the stack before the allocations, which means they will get
picked up by the GC, before popping them back up.

This is a bit inelegant, but the solution to this akwardness is another
IR between us and C anyways, which should come as soon as we get the
basic GC stuff debugged.